requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# Audio processing
pydub>=0.25.1
numpy>=1.24.0

# Machine Learning / AI
openai>=1.0.0
whisper>=1.1.0  # OpenAI Whisper
nemo-toolkit[asr]>=1.20.0  # NeMo pour diarization
omegaconf>=2.3.0

# Document generation
python-docx>=0.8.11
PyPDF2>=3.0.0
pyyaml>=6.0

# HTTP requests
requests>=2.31.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
ruff>=0.0.285
mypy>=1.5.0

# Type stubs
types-requests>=2.31.0
types-PyYAML>=6.0.0
